Rename project to godcell and flatten directory structure
