+++ server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lofi-mmo/server",
+  "name": "@godcell/server",
   "version": "0.1.0",
   "private": true,
   "type": "module",
@@ -9,7 +9,7 @@
     "start": "node dist/index.js"
   },
   "dependencies": {
-    "@lofi-mmo/shared": "*",
+    "@godcell/shared": "*",
     "socket.io": "^4.7.2"
   },
   "devDependencies": {
