+++ shared/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lofi-mmo/shared",
+  "name": "@godcell/shared",
   "version": "0.1.0",
   "private": true,
   "type": "module",
